requirements.txt
setuptools
google-cloud-bigquery
